add rich / loguru compatible logging
