+++ pybs/console/local/__init__.py
@@ -1,0 +1,46 @@
+"""Local commands for the console."""
+import os
+import click as ck
+
+from loguru import logger as log
+from pybs import NAME
+
+@ck.command()
+@ck.argument(
+    "shell", required=False,
+    type=ck.Choice(["bash", "zsh", "fish", "auto"]),
+    default="auto", 
+)
+def completions(
+    shell: str,
+):
+    """Generate shell completion scripts for your shell. 
+    If no shell is specified, it will default to $SHELL."""
+
+    script_name = f"{ck.get_current_context().parent.info_name}"
+    if shell is None or shell == "auto":
+        log.info("No shell specified, defaulting to $SHELL.")
+        shell = os.environ.get("SHELL")
+        log.info(f"Detected shell: {shell}") 
+        shell = shell.split("/")[-1]
+        log.info(f"Extracted shell name: {shell}")  
+    else:
+        log.info(f"Using user-provided shell: {shell}")
+    if shell is None:
+        log.error("No shell detected. Exiting.")
+
+    if shell not in ["bash", "zsh", "fish"]:
+        log.error(f"Unsupported shell: {shell}. Exiting.")
+        return
+    # Use click to generate the completion script
+    script = os.popen(f"_PYBS_COMPLETE={shell}_source {script_name}").read()
+    ck.echo(script)
+
+@ck.command()
+def version():
+    """Show the current version of PyBS."""
+    pkg_name = f"{ck.get_current_context().parent.info_name}"
+    import importlib.metadata
+    version = importlib.metadata.version(pkg_name)
+    ck.echo(f"{NAME} {version}")
+
