+++ pybs/console/remote/__init__.py
@@ -1,0 +1,1 @@
+from pybs.console.remote.code import code
