+++ pyproject.toml
@@ -26,7 +26,7 @@ packages = [
 ]
 
 [project.scripts]
-pybs = "pybs.console:entry_point"
+pybs = "pybs.console.console:entry_point"
